requirements.txt
openai==2.8.0
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.16.3
pyaudio==0.2.14
numpy==1.26.4
pvporcupine==3.0.2
pygame==2.6.1
spotipy==2.24.0
msal==1.30.0
requests==2.32.3
python-dotenv==1.0.1
httpx>=0.27.0
httpcore>=1.0.2

# Build tool (optional, for creating the executable)
pyinstaller==6.10.0

